Test Platform - UI: Remove old code and update question form
