+++ test-platform-ui/src/app/administration/questions/(components)/(question-form)/preview/index.tsx
@@ -1,5 +1,6 @@
 'use effect';
 
+import { TitleWLine } from '@/components/atoms/CustomModules/AutocompleteTags/styles';
 import CustomTextField from '@/components/atoms/CustomModules/CustomTextField';
 import RichTextArea from '@/components/atoms/Editor/richtext';
 import { QuestionLevels, QuestionType } from '@/libs/definitions';
@@ -16,6 +17,7 @@ import {
   RadioGroup,
   Stack,
   Typography,
+  styled,
 } from '@mui/material';
 import { IQuestionInfo } from '../../modifyQuestion';
 
@@ -59,7 +61,24 @@ function getQuestionCategory(type: string) {
   return type;
 }
 
+const ContentBackground = styled(Box)(
+  () => `
+    background-color: white;
+    padding: 15px 20px;
+    border-radius: 10px;
+    border: solid 1px lightgrey;
+  `,
+);
+
 export function QuestionPreview({ values }: IQuestionInterview) {
+  const TitleWDivider = ({ title }: { title: string }) => {
+    return (
+      <TitleWLine variant="body1" className="py-4">
+        <span className="bg-gray-50">{title}</span>
+      </TitleWLine>
+    );
+  };
+
   return (
     <Stack
       gridTemplateColumns={'1fr fit-content(20%) 1fr'}
@@ -67,103 +86,113 @@ export function QuestionPreview({ values }: IQuestionInterview) {
       gap={4}
     >
       <Box>
-        <CustomTextField
-          name="question"
-          customClass="ring-offset-0 w-full pb-4"
-          multiline
-          maxRows={3}
-          isReadOnly={true}
-          onKeyDown={(event) => {
-            if (event.keyCode === 13) {
-              event.preventDefault();
-            }
+        <ContentBackground
+          style={{
+            margin: '20px 0px',
           }}
-        />
-        <Stack gridTemplateColumns={'1fr 1fr'} display={'grid'} width={400}>
-          <Typography
-            className="flex justify-items-center gap-2 text-gray-400"
-            variant="caption"
-            display="block"
-            gutterBottom
-          >
-            <DescriptionOutlinedIcon
-              sx={{
-                width: '16px',
-                height: '16px',
-              }}
+        >
+          <CustomTextField
+            name="question"
+            customClass="ring-offset-0 w-full pb-4"
+            multiline
+            maxRows={3}
+            isReadOnly={true}
+            onKeyDown={(event) => {
+              if (event.keyCode === 13) {
+                event.preventDefault();
+              }
+            }}
+          />
+          <Stack gridTemplateColumns={'1fr 1fr'} display={'grid'} width={400}>
+            <Typography
+              className="flex justify-items-center gap-2 text-gray-400"
+              variant="caption"
+              display="block"
+              gutterBottom
+            >
+              <DescriptionOutlinedIcon
+                sx={{
+                  width: '16px',
+                  height: '16px',
+                }}
+              />
+              {`Type: ${getQuestionType(values.type)}`}
+            </Typography>
+            <Typography
+              className="flex justify-items-center gap-2 text-gray-400"
+              variant="caption"
+              display="block"
+              gutterBottom
+            >
+              <CollectionsBookmarkOutlinedIcon
+                sx={{
+                  width: '16px',
+                  height: '16px',
+                }}
+              />
+              {`Skill: ${getQuestionCategory(values.category)}`}
+            </Typography>
+            <Typography
+              className="flex justify-items-center gap-2 text-gray-400"
+              variant="caption"
+              display="block"
+              gutterBottom
+            >
+              <StackedBarChartOutlinedIcon
+                sx={{
+                  width: '16px',
+                  height: '16px',
+                }}
+              />
+              {`Level: ${getQuestionLevel(values.level)}`}
+            </Typography>
+            <Typography
+              className="flex justify-items-center gap-2 text-gray-400"
+              variant="caption"
+              display="block"
+              gutterBottom
+            >
+              <AccessTimeIcon
+                sx={{
+                  width: '16px',
+                  height: '16px',
+                }}
+              />
+              {millisToMinutesAndSeconds(values.time) || '30s'}
+            </Typography>
+          </Stack>
+          <Box className="pt-6">
+            <RichTextArea
+              key="questionDescriptionPreview"
+              name="description"
+              data={values.description}
+              isReadOnly={true}
             />
-            {`Type: ${getQuestionType(values.type)}`}
-          </Typography>
-          <Typography
-            className="flex justify-items-center gap-2 text-gray-400"
-            variant="caption"
-            display="block"
-            gutterBottom
-          >
-            <CollectionsBookmarkOutlinedIcon
-              sx={{
-                width: '16px',
-                height: '16px',
-              }}
-            />
-            {`Skill: ${getQuestionCategory(values.categories[0])}`}
-          </Typography>
-          <Typography
-            className="flex justify-items-center gap-2 text-gray-400"
-            variant="caption"
-            display="block"
-            gutterBottom
-          >
-            <StackedBarChartOutlinedIcon
-              sx={{
-                width: '16px',
-                height: '16px',
-              }}
-            />
-            {`Level: ${getQuestionLevel(values.level)}`}
-          </Typography>
-          <Typography
-            className="flex justify-items-center gap-2 text-gray-400"
-            variant="caption"
-            display="block"
-            gutterBottom
-          >
-            <AccessTimeIcon
-              sx={{
-                width: '16px',
-                height: '16px',
-              }}
-            />
-            {millisToMinutesAndSeconds(values.time) || '30s'}
-          </Typography>
-        </Stack>
-        <Box className="pt-6">
+          </Box>
+        </ContentBackground>
+        <TitleWDivider title="Explanation" />
+        <ContentBackground>
           <RichTextArea
             key="questionDescriptionPreview"
-            name="description"
-            data={values.description}
+            name="notes"
+            data={values.notes}
             isReadOnly={true}
           />
-        </Box>
-        <Divider className="pb-6">
-          <Typography variant="body2">Explanation</Typography>
-        </Divider>
-        <RichTextArea
-          key="questionDescriptionPreview"
-          name="notes"
-          data={values.notes}
-          isReadOnly={true}
-        />
+        </ContentBackground>
       </Box>
       <Divider orientation="vertical" variant="middle" flexItem />
       <Box>
-        <Divider className="pb-6">
-          <Typography variant="body2">Options</Typography>
-        </Divider>
+        <TitleWDivider title="Options" />
         <RadioGroup
           aria-labelledby="demo-radio-buttons-group-label"
           defaultValue="female"
           name="radio-buttons-group"
+          sx={{
+            backgroundColor: 'white',
+            padding: '10px 20px',
+            borderRadius: '10px',
+            border: 'solid 1px lightgrey',
+          }}
         >
           {values.options?.map((opt, indx: number) => (
             <FormControlLabel
@@ -172,12 +201,17 @@ export function QuestionPreview({ values }: IQuestionInterview) {
                 <Radio
                   key={`opt-${indx}`}
                   value={indx}
-                  disabled
+                  readOnly={true}
                   checked={values.answers.includes(indx)}
                 />
               }
               label={opt}
-              className="py-2 [&>span]:text-sm"
+              className="py-2 [&>span]:text-sm [&_*]:pointer-events-none"
+              sx={{
+                '&~*:hover': {
+                  cursor: 'default',
+                },
+              }}
             />
           ))}
         </RadioGroup>
