--- test-platform-ui/src/app/administration/questions/(components)/(question-form)/questionLevel.tsx
@@ -1,15 +1,0 @@
-import CustomRadioGroup from '@/components/molecules/CustomRadioGroup';
-import { QuestionLevel } from '@/libs/definitions';
-import { ReactElement } from 'react';
-
-const QuestionStandard = (): ReactElement => {
-  return (
-    <CustomRadioGroup
-      label={'Standard'}
-      controlName={'level'}
-      options={QuestionLevel}
-    />
-  );
-};
-
-export default QuestionStandard;
