+++ test-platform-ui/src/app/administration/questions/(edit)/[question_id]/page.tsx
@@ -24,7 +24,8 @@ const EditQuestion = () => {
         id: data?.id,
         question: data?.question,
         description: data?.description,
-        categories: new Array().concat(data?.category),
+        categories: new Array().concat(data?.categories),
+        category: data?.category,
         notes: data?.notes,
         answers: data?.answer,
         options: data?.options,
