+++ test-platform-ui/src/app/administration/questions/page.tsx
@@ -304,12 +304,7 @@ const Page = () => {
         </Box>
       </Box>
       <Divider className="my-10" />
-      <DataTable
-        rows={questionList}
-        columns={columns}
-        // rowHeight={170}
-        loading={loading}
-      />
+      <DataTable rows={questionList} columns={columns} loading={loading} />
       <CustomModal ref={modalRef} title={''}>
         <ModalContent />
       </CustomModal>
