--- test-platform-ui/src/app/administration/questions_v2/(components)/(question-form)/questionFields/questionType.tsx
@@ -1,19 +1,0 @@
-import CustomRadioGroup from '@/components/molecules/CustomRadioGroup';
-import { QuestionTypes } from '@/libs/definitions';
-import { ReactElement } from 'react';
-
-interface IQuestionKind {
-  label?: string;
-}
-
-const QuestionKind = ({ label }: IQuestionKind): ReactElement => {
-  return (
-    <CustomRadioGroup
-      label={label}
-      controlName={'type'}
-      options={QuestionTypes}
-    />
-  );
-};
-
-export default QuestionKind;
