--- test-platform-ui/src/app/administration/questions_v2/(components)/(question-form)/questionFields/skillField.tsx
@@ -1,44 +1,0 @@
-import { IOptions } from '@/components/atoms/CustomModules/AutocompleteTags';
-import AutocompleteQFields from './autocompleteCommonFields';
-
-interface ISkill extends IOptions {
-  value: string;
-  [k: string]: any;
-}
-
-const ArraySkills: ISkill[] = [
-  {
-    name: 'React',
-    value: 'REACT',
-  },
-  {
-    name: 'Javascript',
-    value: 'JAVASCRIPT',
-  },
-  {
-    name: 'HTML',
-    value: 'HTML',
-  },
-  {
-    name: 'CSS',
-    value: 'CSS',
-  },
-  {
-    name: 'Typescript',
-    value: 'TYPESCRIPT',
-  },
-];
-
-export default function AutocompleteSkill({
-  controlName,
-}: {
-  controlName: string;
-}) {
-  return (
-    <AutocompleteQFields
-      controlName={controlName}
-      options={ArraySkills}
-      label="Skill"
-    />
-  );
-}
